mockdata/requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
python-twitter==3.4.2
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
Werkzeug==0.14.1
